Move fp-bindgen generators under feature flag
