+++ example-protocol/Cargo.toml
@@ -5,7 +5,11 @@ authors = ["Fiberplane <info@fiberplane.com>"]
 edition = "2018"
 
 [dependencies]
-fp-bindgen = { path = "../fp-bindgen", features = ["serde-bytes-compat", "time-compat"] }
+fp-bindgen = { path = "../fp-bindgen", features = [
+    "serde-bytes-compat",
+    "time-compat",
+    "generators",
+] }
 pretty_assertions = "0.7"
 serde = { version = "1.0", features = ["derive"] }
 serde_bytes = "0.11"
